backend/medical-codex-chatbot-app-Neo4j-Backend/backend/generate_graphs/requirements.txt
neo4j==5.21.0
requests==2.32.3
python-dotenv==1.0.1
